requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
SQLAlchemy==2.0.27
Werkzeug==3.0.1 